Varios añadidos Checker y Locator
popupmenu listview1
*añadir
*eliminar items
*clear
Locator
*Checkbox avfuck listado a 1 byte
*MultiSelect Listado
Checker
*Checkbox eliminar no funcionales
*Checkbox comprobar fichero de funcionales al terminar en lugar de
hacerlo en tiempo real

+++ uThreadChecker.pas
@@ -38,6 +38,7 @@ end;
 // Procedimiento para ejecutar, matar procesos y errores
 Procedure HChecker.CheckAndKill;
 var
+  TSFuncionales: TStringList;
   i, o: integer;
   TotalFiles, Espera: Integer;
   OFuncionales, Max: Integer;
@@ -45,12 +46,15 @@ var
   hFindW: HWND;
 begin
   OFuncionales:= 0;
+
   if Length(Form1.Edit5.Text) = 0 then
     Max:= 0
   else
     Max:= StrToInt(Form1.Edit5.Text);
+
   Dir:= Form1.EdDir.Text;
   Funcionales:= Dir + '\' + Form1.EdFuncionales.Text;
+
   if Form1.EdEspera.Text = '' then
     Espera:= 750
   else
@@ -77,25 +81,34 @@ begin
       Form1.Estado.Caption:= 'Comprobando fichero: ' + FicheroActual;
       ShellExecute(0, 'open', PChar(RutaCompleta), nil, nil, SW_SHOW);
       Sleep(Espera);
-      if FileExists(Funcionales) then
+
+      if not Form1.ChkRevFinal.Checked then
         begin
-          while FileExists(Funcionales) do
-            DeleteFile(Funcionales);
-          Form1.ListView2.Items.Item[i].SubItems[0]:= 'Sí';
-          Application.ProcessMessages;
-          inc(OFuncionales);
+          if FileExists(Funcionales) then
+          begin
+            while FileExists(Funcionales) do
+              DeleteFile(Funcionales);
+            Form1.ListView2.Items.Item[i].SubItems[0]:= 'Sí';
+            Application.ProcessMessages;
+            inc(OFuncionales);
+          end
+          else
+          begin
+            Form1.ListView2.Items.Item[i].SubItems[0]:= 'No';
+            Application.ProcessMessages;
+          end;
         end
         else
         begin
-          Form1.ListView2.Items.Item[i].SubItems[0]:= 'No';
-          Application.ProcessMessages;
+          Form1.ListView2.Items.Item[i].SubItems[0]:= 'En espera';
+            Application.ProcessMessages;
         end;
 
       if Form1.CheckBox2.Checked then   //Scroll ListView2
         begin
           Form1.ListView2.Items.Item[i].Selected:= True;
           Form1.ListView2.SetFocus;
-          if i > 12 then
+          if i > 10 then
             Form1.ListView2.Scroll(0, Abs(Form1.ListView2.Font.Height));
           Application.ProcessMessages;
         end;
@@ -105,7 +118,36 @@ begin
       PostMessage(hFindW, WM_QUIT, 0, 0);
       KillProcessByName(FicheroActual);
     end;
-  //Le damos unas pasadas por si algún cabrón no se cerró
+
+  if Form1.ChkRevFinal.Checked then
+    try
+    TSFuncionales:= TStringList.Create;
+    try
+      TSFuncionales.LoadFromFile(Funcionales);
+      for I := 0 to Form1.ListView2.Items.Count -1 do
+        for o := 0 to TSFuncionales.Count -1 do
+          begin
+            if pos(TSFuncionales.Strings[o], Form1.Listview2.Items.Item[i].Caption) <> 0 then
+              begin
+                Form1.ListView2.Items.Item[i].SubItems[0]:= 'Sí';
+                inc(OFuncionales);
+                Application.ProcessMessages;
+                Break;
+              end
+              else
+              begin
+                Form1.ListView2.Items.Item[i].SubItems[0]:= 'No';
+                Application.ProcessMessages;
+              end;
+          end;
+    except
+
+    end;
+  finally
+    TSFuncionales.Free;
+  end;
+
+    //Le damos unas pasadas por si algún cabrón no se cerró
   for o := 1 to 5 do
     for i := 0 to TotalFiles -1 do
       begin
@@ -119,6 +161,17 @@ begin
         PostMessage(hFindW, WM_QUIT, 0, 0);
         KillProcessByName(FicheroActual);
     end;
+
+    //Si está marcado, eliminamos las no funcionales
+  if Form1.ChkElimNoF.Checked then
+    for I := 0 to TotalFiles -1 do
+      if Form1.ListView2.Items.Item[i].SubItems[0] = 'No' then
+        begin
+          DeleteFile(Form1.ListView2.Items.Item[i].Caption);
+          Form1.Estado.Caption:= 'Eliminando ficheros no funcionales...';
+          Application.ProcessMessages;
+        end;
+
   if NOT Form1.CheckBox1.Checked then
     Form1.Estado.Caption := 'Proceso completado. Se encontraron ' + IntToStr(OFuncionales) + ' funcionales.'
   else
